Sonar fixer
